Add UI Minimap
Created a UI minimap that both shows the player as a blue arrow, as well as enemies as red arrows. Easy to adjust the viewpoint of the minimaps scene capture camera, so I can adjust the range of the minimap as needed/wanted. Do not yet have friendly NPCs yet, but I do have the sprite prepared for adding them to the minimap as well.